+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+v4.XX.X (Month 2024)
+  - Separate general importer into Acunetix360 & Standard importers
+
 v4.13.0 (July 2024)
   - No changes